backend/python/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
duckduckgo_search==8.0.2
et_xmlfile==2.0.0
filelock==3.18.0
fsspec==2025.3.2
groq==0.25.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
jpype1==1.5.2
lxml==5.4.0
numpy==2.2.5
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pillow==11.2.1
primp==0.15.0
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.12.1
PyMuPDF==1.25.5
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tabula-py==2.10.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
